fix(progress-bar): prevent deletion of incomplete notifications
Add validation in `deleteNotification` method to check if progress is 100% before deletion. If not, a warning notification is sent to inform the user that the task is still in progress. Also, moved `wire:poll.1s.visible` to the inner div to ensure polling only occurs when notifications are visible.

+++ app/Livewire/ProgressBarComponent.php
@@ -3,8 +3,9 @@
 namespace App\Livewire;
 
 use Livewire\Component;
-use App\Services\ProgressNotificationService;
 use Livewire\Attributes\On;
+use Filament\Notifications\Notification;
+use App\Services\ProgressNotificationService;
 
 class ProgressBarComponent extends Component
 {
@@ -40,6 +41,16 @@ class ProgressBarComponent extends Component
     // Método para excluir uma notificação específica
     public function deleteNotification($notificationId)
     {
+        // antes de excluir eu verifico se a progresso esta em 100%
+        if ($this->notifications[$notificationId]['progress'] != 100) {
+            // exclui a notificação
+            // disparo uma notificao que nao pode ser excluida pois ainda nao acabou a tarefa
+            Notification::make()
+                ->title('Não é possível excluir a notificação, ainda não acabou a tarefa!')
+                ->warning()
+                ->send();
+            return;
+        }
         $progressService = app(ProgressNotificationService::class);
         $progressService->deleteNotification($notificationId);
         $this->getNotifications();
